 service/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.31.1
setuptools==80.9.0
sqlparse==0.5.3
